+++ example/views/dummy/subviews/details/index.js
@@ -6,16 +6,50 @@ import Checkbox from 'components/checkbox';
 import Row from 'components/row';
 import FinancesActions from './../../../../actions/finances';
 import Dropdown from 'components/dropdown';
+import DropdownFilter from 'components/dropdown-filter';
+import DropdownFilterAjax from 'components/dropdown-filter-ajax';
 
 class Details extends React.Component {
 
   render() {
     return (
       <Row>
-        <Textbox name="name" value={ this.props.name } onChange={ FinancesActions.financesValueUpdated } validations={ [Presence()] } />
-        <Date name="date_from" value={ this.props.dateFrom } onChange={ FinancesActions.financesValueUpdated } />
-        <Checkbox name="discount" label="Apply Discount?" value={ this.props.discount } onChange={ FinancesActions.financesValueUpdated } />
-        <Dropdown name="accounts" options={ this.props.options } onChange={ FinancesActions.financesValueUpdated } value={ this.props.accounts }/>
+
+        <Textbox
+          name="name"
+          value={ this.props.name }
+          onChange={ FinancesActions.financesValueUpdated }
+          validations={ [Presence()] }
+        />
+
+        <Date
+          name="date_from"
+          value={ this.props.dateFrom }
+          onChange={ FinancesActions.financesValueUpdated }
+        />
+
+        <Checkbox
+          name="discount"
+          label="Apply Discount?"
+          value={ this.props.discount }
+          onChange={ FinancesActions.financesValueUpdated }
+        />
+
+        <Dropdown
+          name="accounts"
+          options={ this.props.options }
+          onChange={ FinancesActions.financesValueUpdated }
+          value={ this.props.accounts }
+        />
+
+        <DropdownFilterAjax
+          name="country"
+          path="/countries"
+          onChange={ FinancesActions.financesCountryUpdated }
+          value={ this.props.countryValue }
+          visibleValue={ this.props.countryVisibleValue }
+        />
+
       </Row>
     );
   }